backend/requirements.txt
# Flask and Web Framework
flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

#Pydantic
pydantic
pydantic-settings

# LangChain and AI
langchain
langchain-openai
langchain-community

# Google Gemini AI (instead of OpenAI)
langchain-google-genai==2.0.5
google-generativeai==0.8.3

# Ollama Integration
langchain-ollama

# Local Embeddings
sentence-transformers
langchain-huggingface
torch

# PDF Processing
pypdf==5.1.0

# Vector Database
faiss-cpu

# Utilities
python-dotenv==1.0.1

# Additional Dependencies
numpy
requests
